fix slot maker where if loopday is larger than working days

+++ database/migrations/2017_11_22_141838_create_dayOffSchedule_table.php
@@ -15,7 +15,7 @@ class CreateDayOffScheduleTable extends Migration
     {
         Schema::create('dayOffSchedule', function (Blueprint $table) {
             $table->increments('id');
-            $table->tinyInteger('slotId');
+            $table->integer('slotId');
             $table->string('date');
             $table->string('description');
             $table->timestamps();
